+++ Tests/NIORedisTests/Commands/HashCommandsTests.swift
@@ -11,7 +11,7 @@ final class HashCommandsTests: XCTestCase {
         do {
             connection = try redis.makeConnection().wait()
         } catch {
-            XCTFail("Failed to create NIORedisConnection!")
+            XCTFail("Failed to create RedisConnection!")
         }
     }
