+++ Tests/NIORedisTests/Commands/StringCommandsTests.swift
@@ -13,7 +13,7 @@ final class StringCommandsTests: XCTestCase {
         do {
             connection = try redis.makeConnection().wait()
         } catch {
-            XCTFail("Failed to create NIORedisConnection!")
+            XCTFail("Failed to create RedisConnection!")
         }
     }
